requirements.txt
Flask==2.0.1
numpy
requests
torch
torchvision
scipy
click
azure-functions